+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.0
 
 - Added commands to sign in and sign out of Azure CLI from the command palette.
+- Added functionality to manage the front-channel logout URL of an application.
 - Added command filters to ensure those that aren't relevant are't shown in the palette.
 - Refactored tree rendering trigger to facilitate better unit testing.
 - Fixed an issue where the wrong error message was passed.